fix: return before attempting to process C/C++ docs
The generated C/C++ docs don't have the same markdown issues as the
markdown docs generated by Sphinx for Python, so if the regex fails to
match any of the problem areas, just return the unprocessed string in
the `parse_content()` function.

This change also moves the location of `main.py` to the `hugoify/`
directory. This allows mounting like `-v "${PWD}/hugoify":"/app/hugoify"`
for local development without clobbering the dependencies installed
by the Dockerfile.

Signed-off-by: Nick Anderegg <nick.anderegg@zymbit.com>