feat: generate custom heading IDs
For permalinks to sections and within the table of contents, Hugo will
automatically generate an ID for the anchor link. Because the full content
of the line can be very long for methods with many parameters, we can
generate our own simplified heading ID to append to each line, which
Hugo will then use.

Signed-off-by: Nick Anderegg <nick.anderegg@zymbit.com>